nuoblog/functions/add_content/requirements.txt
google-cloud-storage==1.13.2